Remove unnecessary type conversion
dct_const_round_shift() takes tran_high_t as parameter and Sqrt2 is of
type tran_high_t. It is unnecessary to convert to int64_t then back
to tran_high_t.

+++ av1/common/av1_inv_txfm1d.c
@@ -1567,7 +1567,7 @@ void av1_iidentity4_c(const int32_t *input, int32_t *output,
                       const int8_t *cos_bit, const int8_t *stage_range) {
   (void)cos_bit;
   for (int i = 0; i < 4; ++i)
-    output[i] = (int32_t)dct_const_round_shift(Sqrt2 * (int64_t)input[i]);
+    output[i] = (int32_t)dct_const_round_shift(Sqrt2 * input[i]);
   range_check(0, input, output, 4, stage_range[0]);
 }
 
@@ -1582,7 +1582,7 @@ void av1_iidentity16_c(const int32_t *input, int32_t *output,
                        const int8_t *cos_bit, const int8_t *stage_range) {
   (void)cos_bit;
   for (int i = 0; i < 16; ++i)
-    output[i] = (int32_t)dct_const_round_shift(Sqrt2 * (int64_t)input[i] * 2);
+    output[i] = (int32_t)dct_const_round_shift(Sqrt2 * input[i] * 2);
   range_check(0, input, output, 16, stage_range[0]);
 }
